Another -Wformat-signedness fix that was missed in dee1c5ac841ef11151ddf92b1ca98be7041f0785.

+++ coregrind/m_gdbserver/remote-utils.c
@@ -1158,7 +1158,7 @@ void prepare_resume_reply (char *buf, char status, unsigned char sig)
          VG_(sprintf) (buf, "%s:%x;",
                        valgrind_stopped_before_syscall ()
                        ? "syscall_entry" : "syscall_return",
-                       valgrind_stopped_by_syscall ());
+                       (UInt)valgrind_stopped_by_syscall ());
          buf += strlen (buf);
       }
